Rename ManagementClient namespace to not duplicate class name
