+++ src/NServiceBus.Transport.RabbitMQ.Tests/FakeHttpClient.cs
@@ -3,14 +3,13 @@
 namespace NServiceBus.Transport.RabbitMQ.Tests
 {
     using System;
-    using System.Net.Http;
     using System.Net;
+    using System.Net.Http;
     using System.Text;
+    using System.Text.Json;
     using System.Threading;
     using System.Threading.Tasks;
-    using ManagementClient;
-    using Queue = ManagementClient.Queue;
-    using System.Text.Json;
+    using NServiceBus.Transport.RabbitMQ.ManagementApi;
 
     class FakeHttpClient
     {
