+++ src/NServiceBus.Transport.RabbitMQ/Administration/ManagementApi/Converters/VersionConverter.cs
@@ -1,6 +1,6 @@
 ﻿#nullable enable
 
-namespace NServiceBus.Transport.RabbitMQ.ManagementClient;
+namespace NServiceBus.Transport.RabbitMQ.ManagementApi;
 
 using System;
 using System.Text.Json;
