+++ src/NServiceBus.Transport.RabbitMQ/Administration/ManagementApi/HttpStatusCodeExtensions.cs
@@ -1,6 +1,6 @@
 ﻿#nullable enable
 
-namespace NServiceBus.Transport.RabbitMQ.ManagementClient;
+namespace NServiceBus.Transport.RabbitMQ.ManagementApi;
 
 using System.Net;
