+++ src/NServiceBus.Transport.RabbitMQ/Administration/ManagementApi/Models/FeatureFlagList.cs
@@ -1,6 +1,6 @@
 ﻿#nullable enable
 
-namespace NServiceBus.Transport.RabbitMQ.ManagementClient;
+namespace NServiceBus.Transport.RabbitMQ.ManagementApi;
 
 using System;
 using System.Collections.Generic;
