+++ src/NServiceBus.Transport.RabbitMQ/Administration/ManagementApi/Models/QueueType.cs
@@ -1,4 +1,4 @@
-﻿namespace NServiceBus.Transport.RabbitMQ.ManagementClient;
+﻿namespace NServiceBus.Transport.RabbitMQ.ManagementApi;
 
 // Note that this is different to `NServiceBus.QueueType` which lists the types of queues supported
 // by the NServiceBus transport, which doesn't include the `Stream` value
