+++ src/NServiceBus.Transport.RabbitMQ/Administration/ManagementApi/Response.cs
@@ -1,6 +1,6 @@
 ﻿#nullable enable
 
-namespace NServiceBus.Transport.RabbitMQ.ManagementClient;
+namespace NServiceBus.Transport.RabbitMQ.ManagementApi;
 
 using System.Diagnostics.CodeAnalysis;
 using System.Net;
